Add Production ETL Parcels Routine to Github Repo
This is the initial commit to transition to version control for the
Town of Prosper GIS automations. This script routine is for an ETL for
grabbing Collin Central Appraisal District (CCAD) and Denton Central
Appraisal District (DCAD) parcels, amending their schema, and pushing to
the production environment. This routine is meant to run nightly to
capture any changes made by either entity.

Tags: CCAD, DCAD, parcels, ETL

Authored-by: Jordan Carmona <12204584+FeralCatColonist@users.noreply.github.com>