Add Production ETL Addresses, Roads to Github Repo
This is the initial commit to transition to version control for the
Town of Prosper GIS automations. This script routine is for an ETL for
grabbing various information from the regional emergency communications
districts NCT-911 and DenCo911. This script grabs the data amends their
schema, performs various cleaning functions in terms of data and
geographic area, and pushes it to the production environment. This
routine is meant to run nightly to capture any changes made by NCT-911
partners and the DenCo911 portion is run monthly as data is available
for download.

Tags: NCT-911, DenCo911, roads, addresses, ETL

Authored-by: Jordan Carmona <12204584+FeralCatColonist@users.noreply.github.com>