Add Production SQL Backup and Restore for ViewerDB
This is the initial commit to transition to version control for the
Town of Prosper GIS automations. This script routine is for backing up
the [Production] GIS database and restoring it to the [Viewer] GIS
Database for use as a datasource for the ArcGIS Portal environment. It
uses SQLCMD inside of a SQL Server Agent Job to transfer the data
nightly via a BACKUP and RESTORE. 

Tags: SQL, Datastore

Authored-by: Jordan Carmona <12204584+FeralCatColonist@users.noreply.github.com>