Drop Database Maintenance from Main Script
Story:
The database maintenance routine utilized before the GIS team had a lot
of members was a bit haphazard and did not contribute to effective state
control of the database. As such it will be removed from the ETL and
implemented into a separate database maintenance routine to be run after
the other ETL processes have finished.

Tags:
database maintenance,

Authored-by: Jordan Carmona <12204584+FeralCatColonist@users.noreply.github.com>