Add Production Tyler EPL Refresh to Github Repo
This is the initial commit to transition to version control for the
Town of Prosper GIS automations. This script routine is for updating
the enterprise geodatabase that holds Tyler EPL data with fresh data
from the production environment. The Tyler EPL data requires a
WGS84 projection which is why we pushed it into its own enterprise
geodatabase.