Fix Drop Fields Error in data_cleaner
Story:
There was a schema change on the part of DenCo911 that broke the script.
This could have been avoided if the sedf.drop() had been written with
the parameter `errors = "ignore"`; as this is a drop command, that would
be an appropriate change to make with minimal consequences. This commit
adds that parameter.

Tags:
bug,

Authored-by: Jordan Carmona <12204584+FeralCatColonist@users.noreply.github.com>