Add DB User Re-map to GISVIEW RESTORE
Story:
This change was made as an insurance policy for the ArcGIS Online
Collaboration. Generally, that collaboration will be done via replica
from the Production DB but in case we needed an alternative method I am
now adding this to the user re-mapping that is done for the nightly
RESTORE procedure.

Tags:
collaboration, db user, SQL

Authored-by: Jordan Carmona <12204584+FeralCatColonist@users.noreply.github.com>